feat/#18: 상품 상세 페이지 이동 시 상품 ID 전달 로직 추가

+++ app/src/main/java/com/sopt/bunjang/presentation/home/navigation/HomeNavigation.kt
@@ -16,7 +16,7 @@ fun NavController.navigateHome(navOptions: NavOptions? = null) {
 fun NavGraphBuilder.homeGraph(
     paddingValues: PaddingValues,
     navigateUp: () -> Unit,
-    navigateToProductDetail: () -> Unit
+    navigateToProductDetail: (Long) -> Unit
 ) {
     composable<Home> {
         HomeRoute(
